Refactored (consolidated) redundant code
